Conflicts:
	app/src/main/java/hsleiden/ikpmd3/levelStateManager/Level1Activity.java
	app/src/main/java/hsleiden/ikpmd3/player/Player.java

+++ app/src/main/java/hsleiden/ikpmd3/map/Tile.java
@@ -19,7 +19,7 @@ public class Tile {
 		this.image = image;
 	}
 
-	public void moveTileLeft(int xToMove)
+	public void moveTileLeft(float xToMove)
 	{
 		this.x -= xToMove;
 	}
